Numerical Simulations
test.m: create a sparse binary matrix, symmetrize it and check how the estimation error of the smallest eigenvalue varies with submatrix size
lower_bound.m: out of n, choose cn indices (0<c<1) and set the submatrix corresponding to these cn indices to -1 and set the rest of the entries to 0. This case achieves the 1/eps^2 lower bound and I think can be put in the simulations section.